requirements.txt
# Riosanatea 宛名印刷ソフト - 必要なライブラリ
# Ubuntu 22.04 / 24.04 対応

# GUI フレームワーク
wxPython>=4.1.0

# 画像処理
Pillow>=9.0.0
